fix:Repair star acquisition failure
