+++ .github/workflows/cron-deploy-website.yml
@@ -39,6 +39,7 @@ jobs:
                   export NODE_OPTIONS=--max-old-space-size=8192
                   yarn
                   node ./scripts/update_search_hash.js
+                  node ./scripts/update_github_info.js
                   yarn build
                   touch build/.dummy
                   cp .asf-site.yaml ./build/.asf.yaml
