+++ src/constant/github.data.ts
@@ -1,0 +1,2 @@
+// The number of stars will be automatically updated with each deployment
+export const STAR_COUNT = '13.1';
